+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+
+## [1.1.0]
 ### Added
 - Postman / Newman integration tests to ensure giuseppe works with the given express version.
 - Possibility to add multiple route decorators to the same function.
@@ -93,7 +96,8 @@ No news :-)
 - First version of giuseppe.
 - Adding base code and travis-CI.
 
-[Unreleased]: https://github.com/smartive/giuseppe/compare/v1.0.0...master
+[Unreleased]: https://github.com/smartive/giuseppe/compare/v1.1.0...master
+[1.1.0]: https://github.com/smartive/giuseppe/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/smartive/giuseppe/compare/v1.0.0-rc.1...v1.0.0
 [1.0.0-rc.1]: https://github.com/smartive/giuseppe/compare/v0.6.1...v1.0.0-rc.1
 [0.6.1]: https://github.com/smartive/giuseppe/compare/v0.6.0...v0.6.1